+++ pkg/kube/translation/annotations/plugins/authorization.go
@@ -12,9 +12,10 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package annotations
+package plugins
 
 import (
+	"github.com/apache/apisix-ingress-controller/pkg/kube/translation"
 	apisixv1 "github.com/apache/apisix-ingress-controller/pkg/types/apisix/v1"
 )
 
@@ -35,8 +36,8 @@ func (b *basicAuth) PluginName() string {
 	return "basic-auth"
 }
 
-func (b *basicAuth) Handle(e Extractor) (interface{}, error) {
-	if e.GetStringAnnotation(_authType) != "basicAuth" {
+func (b *basicAuth) Handle(ing *translation.Ingress) (interface{}, error) {
+	if GetStringAnnotation(_authType) != "basicAuth" {
 		return nil, nil
 	}
 	plugin := apisixv1.BasicAuthConfig{}
@@ -55,8 +56,8 @@ func (k *keyAuth) PluginName() string {
 	return "key-auth"
 }
 
-func (k *keyAuth) Handle(e Extractor) (interface{}, error) {
-	if e.GetStringAnnotation(_authType) != "keyAuth" {
+func (k *keyAuth) Handle(ing *translation.Ingress) (interface{}, error) {
+	if GetStringAnnotation(_authType, ing) != "keyAuth" {
 		return nil, nil
 	}
 	plugin := apisixv1.KeyAuthConfig{}
