+++ pkg/kube/translation/annotations/plugins/cors.go
@@ -12,35 +12,38 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package annotations
+package plugins
 
 import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-
 	apisixv1 "github.com/apache/apisix-ingress-controller/pkg/types/apisix/v1"
 )
 
-func TestCorsHandler(t *testing.T) {
-	annotations := map[string]string{
-		_enableCors:       "true",
-		_corsAllowHeaders: "abc,def",
-		_corsAllowOrigin:  "https://a.com",
-		_corsAllowMethods: "GET,HEAD",
+const (
+	_enableCors       = AnnotationsPrefix + "enable-cors"
+	_corsAllowOrigin  = AnnotationsPrefix + "cors-allow-origin"
+	_corsAllowHeaders = AnnotationsPrefix + "cors-allow-headers"
+	_corsAllowMethods = AnnotationsPrefix + "cors-allow-methods"
+)
+
+type cors struct{}
+
+// NewCorsHandler creates a handler to convert annotations about
+// CORS to APISIX cors plugin.
+func NewCorsHandler() Handler {
+	return &cors{}
+}
+
+func (c *cors) PluginName() string {
+	return "cors"
+}
+
+func (c *cors) Handle(e Extractor) (interface{}, error) {
+	if !e.GetBoolAnnotation(_enableCors) {
+		return nil, nil
 	}
-	p := NewCorsHandler()
-	out, err := p.Handle(NewExtractor(annotations))
-	assert.Nil(t, err, "checking given error")
-	config := out.(*apisixv1.CorsConfig)
-	assert.Equal(t, "abc,def", config.AllowHeaders)
-	assert.Equal(t, "https://a.com", config.AllowOrigins)
-	assert.Equal(t, "GET,HEAD", config.AllowMethods)
-
-	assert.Equal(t, "cors", p.PluginName())
-
-	annotations[_enableCors] = "false"
-	out, err = p.Handle(NewExtractor(annotations))
-	assert.Nil(t, err, "checking given error")
-	assert.Nil(t, out, "checking given output")
+	return &apisixv1.CorsConfig{
+		AllowOrigins: e.GetStringAnnotation(_corsAllowOrigin),
+		AllowMethods: e.GetStringAnnotation(_corsAllowMethods),
+		AllowHeaders: e.GetStringAnnotation(_corsAllowHeaders),
+	}, nil
 }
