+++ pkg/kube/translation/annotations/plugins/cors_test.go
@@ -12,7 +12,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package annotations
+package plugins
 
 import (
 	"testing"
@@ -22,33 +22,25 @@ import (
 	apisixv1 "github.com/apache/apisix-ingress-controller/pkg/types/apisix/v1"
 )
 
-func TestIPRestrictionHandler(t *testing.T) {
+func TestCorsHandler(t *testing.T) {
 	annotations := map[string]string{
-		_allowlistSourceRange: "10.2.2.2,192.168.0.0/16",
+		_enableCors:       "true",
+		_corsAllowHeaders: "abc,def",
+		_corsAllowOrigin:  "https://a.com",
+		_corsAllowMethods: "GET,HEAD",
 	}
-	p := NewIPRestrictionHandler()
+	p := NewCorsHandler()
 	out, err := p.Handle(NewExtractor(annotations))
 	assert.Nil(t, err, "checking given error")
-	config := out.(*apisixv1.IPRestrictConfig)
-	assert.Len(t, config.Allowlist, 2, "checking size of white list")
-	assert.Equal(t, "10.2.2.2", config.Allowlist[0])
-	assert.Equal(t, "192.168.0.0/16", config.Allowlist[1])
-	assert.Equal(t, "ip-restriction", p.PluginName())
+	config := out.(*apisixv1.CorsConfig)
+	assert.Equal(t, "abc,def", config.AllowHeaders)
+	assert.Equal(t, "https://a.com", config.AllowOrigins)
+	assert.Equal(t, "GET,HEAD", config.AllowMethods)
 
-	annotations[_blocklistSourceRange] = "172.17.0.0/16,127.0.0.1"
+	assert.Equal(t, "cors", p.PluginName())
+
+	annotations[_enableCors] = "false"
 	out, err = p.Handle(NewExtractor(annotations))
 	assert.Nil(t, err, "checking given error")
-	config = out.(*apisixv1.IPRestrictConfig)
-	assert.Len(t, config.Allowlist, 2, "checking size of white list")
-	assert.Equal(t, "10.2.2.2", config.Allowlist[0])
-	assert.Equal(t, "192.168.0.0/16", config.Allowlist[1])
-	assert.Len(t, config.Blocklist, 2, "checking size of black list")
-	assert.Equal(t, "172.17.0.0/16", config.Blocklist[0])
-	assert.Equal(t, "127.0.0.1", config.Blocklist[1])
-
-	delete(annotations, _allowlistSourceRange)
-	delete(annotations, _blocklistSourceRange)
-	out, err = p.Handle(NewExtractor(annotations))
-	assert.Nil(t, err, "checking given error")
-	assert.Nil(t, out, "checking the given ip-restrction plugin config")
+	assert.Nil(t, out, "checking given output")
 }
