+++ pkg/kube/translation/annotations/plugins/forward_auth.go
@@ -12,7 +12,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package annotations
+package plugins
 
 import (
 	apisixv1 "github.com/apache/apisix-ingress-controller/pkg/types/apisix/v1"
